Remove execHandler
